refactor: MemberRepository 네이밍, 패키지 변경
