+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/meeting/JpaMeetingRepository.java
@@ -1,11 +1,11 @@
-package com.mobble.mobbleserver.refactor.meeting.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.meeting;
 
 import com.mobble.mobbleserver.infrastructure.web.meeting.Meeting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MeetingRepository extends JpaRepository<Meeting, Long> {
+public interface JpaMeetingRepository extends JpaRepository<Meeting, Long> {
     //    List<Meeting> findMeetingsByClubId(Long clubId);
     List<Meeting> findByClubMember_Club_Id(Long clubId);
 }
